api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
earthengine-api==0.1.374
geemap==0.28.2
pandas>=2.2.0
numpy>=1.25.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
google-auth==2.23.4
pydantic==2.4.2